+++ sphinx/application/common/wrappers/wrapper-podcast/src/main/java/chat/sphinx/wrapper_podcast/FeedRecommendation.kt
@@ -4,6 +4,7 @@ import java.util.*
 
 data class FeedRecommendation(
     val id: String,
+    val pubKey: String?,
     val feedType: String,
     val description: String,
     val smallImageUrl: String?,