+++ sphinx/application/sphinx/src/main/java/chat/sphinx/di/NetworkModule.kt
@@ -12,6 +12,7 @@ import chat.sphinx.concept_network_query_action_track.NetworkQueryActionTrack
 import chat.sphinx.concept_network_query_chat.NetworkQueryChat
 import chat.sphinx.concept_network_query_contact.NetworkQueryContact
 import chat.sphinx.concept_network_query_crypter.NetworkQueryCrypter
+import chat.sphinx.concept_network_query_discover_tribes.NetworkQueryDiscoverTribes
 import chat.sphinx.concept_network_query_invite.NetworkQueryInvite
 import chat.sphinx.concept_network_query_lightning.NetworkQueryLightning
 import chat.sphinx.concept_network_query_meme_server.NetworkQueryMemeServer
@@ -34,6 +35,7 @@ import chat.sphinx.feature_network_query_action_track.NetworkQueryActionTrackImp
 import chat.sphinx.feature_network_query_chat.NetworkQueryChatImpl
 import chat.sphinx.feature_network_query_contact.NetworkQueryContactImpl
 import chat.sphinx.feature_network_query_crypter.NetworkQueryCrypterImpl
+import chat.sphinx.feature_network_query_discover_tribes.NetworkQueryDiscoverTribesImpl
 import chat.sphinx.feature_network_query_invite.NetworkQueryInviteImpl
 import chat.sphinx.feature_network_query_lightning.NetworkQueryLightningImpl
 import chat.sphinx.feature_network_query_meme_server.NetworkQueryMemeServerImpl
@@ -453,4 +455,17 @@ object NetworkModule {
         networkQueryActionTrackImpl: NetworkQueryActionTrackImpl
     ): NetworkQueryActionTrack =
         networkQueryActionTrackImpl
+
+    @Provides
+    fun provideNetworkQueryDiscoverTribesImpl(
+        networkRelayCall: NetworkRelayCall
+    ): NetworkQueryDiscoverTribesImpl =
+        NetworkQueryDiscoverTribesImpl(networkRelayCall)
+
+    @Provides
+    fun provideNetworkQueryDiscoverTribes(
+        networkQueryDiscoverTribesImpl: NetworkQueryDiscoverTribesImpl
+    ): NetworkQueryDiscoverTribes =
+        networkQueryDiscoverTribesImpl
+
 }