+++ sphinx/screens-detail/common-player/src/main/java/chat/sphinx/common_player/ui/CommonPlayerScreenSideEffect.kt
@@ -37,5 +37,15 @@ sealed class CommonPlayerScreenSideEffect: SideEffect<Context>() {
             override val stringRes: Int
                 get() = R.string.error_loading_recommendations
         }
+
+        object BalanceTooLow : Notify() {
+            override val stringRes: Int
+                get() = R.string.balance_too_low
+        }
+
+        object BoostAmountTooLow : Notify() {
+            override val stringRes: Int
+                get() = R.string.boost_amount_too_low
+        }
     }
 }