+++ sphinx/screens-detail/join-tribe/join-tribe/src/main/java/chat/sphinx/join_tribe/navigation/JoinTribeNavigator.kt
@@ -9,6 +9,7 @@ abstract class JoinTribeNavigator(
     detailNavigationDriver: BaseNavigationDriver<NavController>
 ): Navigator<NavController>(detailNavigationDriver) {
     abstract suspend fun closeDetailScreen()
+
     suspend fun popBackStack() {
         navigationDriver.submitNavigationRequest(PopBackStack())
     }